// frontend/src/app/shared/api/model/get-me200-response.ts
/**
 * Food Driver API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetMe200ResponseStreet } from './get-me200-response-street';
import { UserRole } from './user-role';
import { GetMe200ResponsePostalCode } from './get-me200-response-postal-code';


export interface GetMe200Response { 
  [key: string]: any | any;


    /**
     * Unikalny identyfikator
     */
    idUser: number;
    name: string;
    surname: string;
    email: string;
    street: GetMe200ResponseStreet;
    building: GetMe200ResponseStreet;
    apartment: GetMe200ResponseStreet;
    postalCode: GetMe200ResponsePostalCode;
    city: GetMe200ResponseStreet;
    role: UserRole;
    created: string;
}
export namespace GetMe200Response {
}