// frontend/src/app/shared/api/model/get-me-restaurant200-response-menu-inner.ts
/**
 * Food Driver API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface GetMeRestaurant200ResponseMenuInner {
  [key: string]: any | any;

  /**
   * Unikalny identyfikator
   */
  idMenuItem: number;
  name: string;
  description: string;
  price: string;
  image: string;
  idImage?: string;
}